+++ src/components/DashboardLayout/index.tsx
@@ -1,16 +1,15 @@
-import React, {FC, ReactNode, useEffect, useRef} from 'react';
+import React, {FC, useEffect, useRef} from 'react';
 import {useLocation} from 'react-router-dom';
 
-import {BreadcrumbMenu} from 'components';
+import {BreadcrumbMenu, DocsMenuItems} from 'components';
 import './DashboardLayout.scss';
 
 interface ComponentProps {
-  menuItems: ReactNode;
   pageName: string;
   sectionName: string;
 }
 
-const DashboardLayout: FC<ComponentProps> = ({children, menuItems, pageName, sectionName}) => {
+const DashboardLayout: FC<ComponentProps> = ({children, pageName, sectionName}) => {
   const {pathname} = useLocation();
   const rightDiv = useRef<HTMLDivElement>(null);
 
@@ -21,9 +20,11 @@ const DashboardLayout: FC<ComponentProps> = ({children, menuItems, pageName, sec
   return (
     <div className="DashboardLayout">
       <div className="DashboardLayout__breadcrumbs">
-        <BreadcrumbMenu menuItems={menuItems} pageName={pageName} sectionName={sectionName} />
+        <BreadcrumbMenu menuItems={<DocsMenuItems />} pageName={pageName} sectionName={sectionName} />
       </div>
-      <div className="DashboardLayout__left-menu">{menuItems}</div>
+      <div className="DashboardLayout__left-menu">
+        <DocsMenuItems />
+      </div>
       <div className="DashboardLayout__main-content" ref={rightDiv}>
         {children}
       </div>
