+++ src/components/TopNav/TopNavLogo/TopNavLogo.scss
@@ -1,0 +1,20 @@
+.TopNavLogo {
+  align-items: center;
+  background-color: transparent;
+  border-right: none;
+  display: flex;
+  font-family: var(--font-family-mono);
+  font-size: 16px;
+  font-weight: 700;
+  margin-left: 24px;
+  padding: 8px 0;
+
+  &__img {
+    height: 24px;
+    margin-right: 6px;
+  }
+
+  @media (max-width: 768px) {
+    margin-left: 12px;
+  }
+}
