+++ src/containers/BankApi/index.tsx
@@ -1,8 +1,8 @@
 import React, {FC, useMemo} from 'react';
 import {Redirect, useParams} from 'react-router-dom';
 
-import {ApiMenuItems, DashboardLayout, Pagination} from 'components';
-import {bankApiNavigationData} from 'components/ApiMenuItems';
+import {DashboardLayout, Pagination} from 'components';
+import {bankApiNavigationData} from 'components/DocsMenuItems';
 import NodeApiConnectionRequests from 'containers/NodeApi/NodeApiConnectionRequests';
 import {PageData, PageDataObject} from 'types/page-data';
 
@@ -78,7 +78,7 @@ const BankApi: FC = () => {
   const {content, name} = useMemo(() => getPageData(chapter), [chapter]);
 
   return (
-    <DashboardLayout menuItems={<ApiMenuItems />} pageName={name} sectionName="Bank API">
+    <DashboardLayout pageName={name} sectionName="Bank API">
       {content}
       <Pagination navigationData={bankApiNavigationData} />
     </DashboardLayout>
