+++ src/containers/DeploymentGuide/index.tsx
@@ -2,7 +2,6 @@ import React, {FC, useMemo} from 'react';
 import {Redirect, useParams} from 'react-router-dom';
 
 import {DashboardLayout} from 'components';
-import GuideMenuItems from 'components/GuideMenuItems';
 import {PageData, PageDataObject} from 'types/page-data';
 
 import DeploymentGuideBank from './DeploymentGuideBank';
@@ -33,7 +32,7 @@ const DeploymentGuide: FC = () => {
   const {content, name} = useMemo(() => getPageData(chapter), [chapter]);
 
   return (
-    <DashboardLayout menuItems={<GuideMenuItems />} pageName={name} sectionName="Deployment Guide">
+    <DashboardLayout pageName={name} sectionName="Deployment Guide">
       {content}
     </DashboardLayout>
   );
