+++ src/containers/StyleGuide/index.tsx
@@ -2,7 +2,6 @@ import React, {FC, useMemo} from 'react';
 import {Redirect, useParams} from 'react-router-dom';
 
 import {DashboardLayout} from 'components';
-import GuideMenuItems from 'components/GuideMenuItems';
 import {PageData, PageDataObject} from 'types/page-data';
 
 import StyleGuideCss from './StyleGuideCss';
@@ -33,7 +32,7 @@ const StyleGuide: FC = () => {
   const {content, name} = useMemo(() => getPageData(chapter), [chapter]);
 
   return (
-    <DashboardLayout menuItems={<GuideMenuItems />} pageName={name} sectionName="Style Guide">
+    <DashboardLayout pageName={name} sectionName="Style Guide">
       {content}
     </DashboardLayout>
   );
